refactored postgres schema builder
