+++ core-api/index.d.ts
@@ -33,3 +33,4 @@ export { Action, classBackendMethodsArray, serverActionField } from "./src/serve
 export { actionInfo } from "./src/server-action";
 export { registerActionsOnServer } from './server/register-actions-on-server';
 export { registerEntitiesOnServer } from './server/register-entities-on-server';
+/** db implementation required */
