+++ projects/core/index.ts
@@ -70,6 +70,9 @@ export {
 export {
     SqlDatabase
 } from './src/data-providers/sql-database';//V
+
+
+
 export { JsonDataProvider, JsonEntityStorage } from './src/data-providers/json-data-provider';//V
 
 //export * from './src/data-api'; //reconsider if to make internal
@@ -116,3 +119,7 @@ export { Action, classBackendMethodsArray, serverActionField } from "./src/serve
 export { actionInfo } from "./src/server-action";
 export { registerActionsOnServer } from './server/register-actions-on-server';
 export { registerEntitiesOnServer } from './server/register-entities-on-server';
+
+
+
+/** db implementation required */
