+++ projects/core/postgres/index.ts
@@ -1,1 +1,2 @@
 export * from './postgres-data-provider';
+export * from './schema-builder';
